requirements.txt
numpy==2.3.3
pandas==2.3.2
requests==2.32.5
python-dotenv==1.1.1
pyarrow==21.0.0
pytest==8.4.2
python-binance==1.0.29
aiohttp==3.12.15
websockets==15.0.1
pycryptodome==3.23.0
dateparser==1.2.2
tzlocal==5.3.1
regex==2025.9.18
attrs==25.3.0
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
frozenlist==1.7.0
multidict==6.6.4
propcache==0.3.2
yarl==1.20.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
charset_normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.8.3
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
six==1.17.0
colorama==0.4.6
iniconfig==2.1.0
